Fixed the bin codec: 	- the int1/int2/int4 functions were removed because they were prone to errors. 	- new functions for char,uchar,short,ushort,int,uint,time_t were introduced
Because of the bin codec fix many data structures were changed and also optimized for space.
Lots of changes in the code due to changes in the data structures.

Introduced persistency for P-CSCF subscriptions to reg.

+++ ser_ims/branches/01_persistency/modules/scscf/registration.c
@@ -1182,7 +1182,7 @@ void reg_await_timer(unsigned int ticks, void* param)
 			av = aud->head;
 			while(av){
 				LOG(L_DBG,"DBG:"M_NAME":reg_await_timer: .. AV %4d - %d Exp %3d  %p\n",
-					av->item_number,av->status,av->expires,av);
+					av->item_number,av->status,(int)av->expires,av);
 				av_next = av->next;
 				if ((av->status == AUTH_VECTOR_USELESS || 
 					av->status == AUTH_VECTOR_SENT) && av->expires<ticks)
